Moving logic to python
